+++ integrationTests/cypress/support/step_definitions/legal_urls.js
@@ -1,8 +1,8 @@
 const {
-    Then, Given
+    Then
 } = require("cypress-cucumber-preprocessor/steps");
 
-Then(/^a footer with three legal links is displayed$/, function () {
+Then("a footer with three legal links is displayed", function () {
     cy.get('footer').find('a').should('have.length', 3)
 
     cy.get('footer').get('[data-testId=footer-link-terms-of-service]').contains('Terms of Service').should('have.attr', 'href', Cypress.env("TermsOfServiceURL"))
@@ -11,7 +11,7 @@ Then(/^a footer with three legal links is displayed$/, function () {
 });
 
 
-Then(/^the legal links in the footer are translated to german$/, function () {
+Then("the legal links in the footer are translated to German", function () {
     cy.get('footer').find('a').should('have.length', 3)
 
     cy.get('footer').get('[data-testId=footer-link-terms-of-service]').contains('Nutzungsbedingungen')
@@ -19,6 +19,6 @@ Then(/^the legal links in the footer are translated to german$/, function () {
     cy.get('footer').get('[data-testId=footer-link-privacy-policy]').contains('Datenschutzerklärung')
 });
 
-Then(/^the forgot password button is translated to german$/, function () {
+Then("the forgot password button is translated to German", function () {
     cy.get('button[data-testid="login-forgot-password-button"]').contains('Passwort vergessen?')
 });