+++ integrationTests/cypress/support/step_definitions/login.js
@@ -5,16 +5,16 @@ const {
 } = require("cypress-cucumber-preprocessor/steps");
 const env = require('@cloudogu/dogu-integration-test-library/lib/environment_variables')
 
-Given(/^the browser shows the CAS login page$/, function () {
+Given("the browser shows the CAS login page", function () {
     cy.visit("/cas/login")
     cy.clickWarpMenuCheckboxIfPossible()
 });
 
-Given(/^the user is currently not logged in the CES$/, function () {
+Given("the user is currently not logged in the CES", function () {
     cy.visit("/cas/logout")
 });
 
-Given(/^the browser shows the CAS login page and the browser language is set to german$/, function () {
+Given("the browser shows the CAS login page and the browser language is set to German", function () {
     cy.visit("/cas/login", {
         onBeforeLoad(window) {
             Object.defineProperty(window.navigator, 'language', {value: 'de-DE'});
@@ -28,17 +28,17 @@ Given(/^the browser shows the CAS login page and the browser language is set to 
     cy.clickWarpMenuCheckboxIfPossible()
 })
 
-When(/^the user logs into the CES with the admin credentials$/, function () {
+When("the user logs into the CES with the admin credentials", function () {
     cy.clickWarpMenuCheckboxIfPossible()
     cy.login(env.GetAdminUsername(), env.GetAdminPassword());
 });
 
-When(/^When the user logs into the CES with the admin credentials, writing the username in capital letters$/, function () {
+When("When the user logs into the CES with the admin credentials, writing the username in capital letters", function () {
     cy.clickWarpMenuCheckboxIfPossible()
     cy.login(env.GetAdminUsername().toUpperCase(), env.GetAdminPassword());
 });
 
-Then(/^CAS shows the profile page of the user with the user ID from LDAP entry$/, function () {
+Then("CAS shows the profile page of the user with the user ID from LDAP entry", function () {
     cy.get('h2[data-testid=login-header]').contains("Log In Successful")
     cy.get(':nth-child(2) > strong').contains(env.GetAdminUsername())
 });