+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- Fixed german translation on login page.
+- Fixed german translation on login page (#138).
 
 ## [v6.3.7-4] - 2022-03-29
 ### Changed